Split package: delete 'ca.uhn.fhir.rest.server' package from hapi-fhir-server-empi
